+++ crates/ruma-common/src/identifiers.rs
@@ -39,7 +39,7 @@ pub use self::{
     room_version_id::RoomVersionId,
     server_name::ServerName,
     server_signing_key_version::ServerSigningKeyVersion,
-    session_id::{OwnedSessionId, SessionId},
+    session_id::SessionId,
     signatures::{
         CrossSigningOrDeviceSignatures, DeviceSignatures, EntitySignatures, ServerSignatures,
         Signatures,
@@ -160,7 +160,7 @@ macro_rules! device_id_ref {
 pub mod __private_macros {
     pub use ruma_macros::{
         base64_public_key, event_id, mxc_uri, room_alias_id, room_id, room_version_id, server_name,
-        server_signing_key_version, user_id,
+        server_signing_key_version, session_id, user_id,
     };
 
     #[cfg(feature = "unstable-identifier-ref-macros")]
@@ -228,6 +228,9 @@ pub mod __private_macros {
         pub static SERVER_SIGNING_KEY_VERSION_INTERNER: LazyLock<
             IdInterner<crate::ServerSigningKeyVersion>,
         > = LazyLock::new(IdInterner::new);
+
+        pub static SESSION_ID_INTERNER: LazyLock<IdInterner<crate::SessionId>> =
+            LazyLock::new(IdInterner::new);
     }
 }
 
@@ -367,21 +370,23 @@ macro_rules! server_name_ref {
 /// Compile-time checked [`SessionId`] construction.
 #[macro_export]
 macro_rules! session_id {
-    ($s:literal) => {{
-        const SESSION_ID: &$crate::SessionId = match $crate::SessionId::_priv_const_new($s) {
-            Ok(id) => id,
-            Err(e) => panic!("{}", e),
-        };
-
-        SESSION_ID
-    }};
+    ($s:literal) => {{ $crate::__private_macros::session_id!($crate, $s) }};
 }
 
-/// Compile-time checked [`OwnedSessionId`] construction.
+/// Compile-time checked `&'static SessionId` construction.
+///
+/// This macro is a helper to ease the transition after the change of [`SessionId`] from a
+/// dynamically sized type to an owned type. It has the side effect of interning and leaking the
+/// identifier so it SHOULD NOT be used in code that runs in production.
+///
+/// This is behind the `unstable-identifier-ref-macros` cargo feature to allow us to remove this
+/// macro at any time without it being a breaking change.
 #[macro_export]
-macro_rules! owned_session_id {
+#[cfg(feature = "unstable-identifier-ref-macros")]
+macro_rules! session_id_ref {
     ($s:literal) => {
-        $crate::session_id!($s).to_owned()
+        $crate::__private_macros::id_interner::SESSION_ID_INTERNER
+            .get_or_insert_with($s, || $crate::session_id!($s))
     };
 }